requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap4==2024.1
dateparser==1.2.0
Django==5.0.7
django-crispy-forms==2.3
django-easy-validator==1.7.0
django-environ==0.11.2
django-js-routes==0.2.0
django-rulebase==0.1.3
django-vite==3.0.4
django_neomodel==0.2.0
dnspython==2.6.1
Faker==27.0.0
idna==3.7
inertia-django==0.6.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
neo4j==5.19.0
neomodel==5.3.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
isort